Cari Sil

+++ IlacDepoStok/FormDepo.cs
@@ -116,11 +116,13 @@ namespace IlacDepoStok
 
         private void btnSil_Click(object sender, EventArgs e)
         {
-            DialogResult dr = MessageBox.Show("Seçili Depo'yu Silmek İstediğinizden Emin misiniz?", "Silme Onayı", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-            if(dr.")
-            int depoId=((DepoModel)(lstBoxDepolar.SelectedItem)).id;
-            SqliteDataAccess.DeleteDepo(depoId);
-            
+            DialogResult dialog = new DialogResult();
+            dialog = MessageBox.Show("Seçili Depo'yu Silmek İstediğinizden Emin misiniz?", "Silme Onayı", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (dialog == DialogResult.Yes)
+            {
+                int depoId = ((DepoModel)(lstBoxDepolar.SelectedItem)).id;
+                SqliteDataAccess.DeleteDepo(depoId);
+            }
             loadDepo();
             btnKaydet.Enabled = false;
         }
